pyproject.toml
[project]
name = "aigovern-warden"
version = "0.1.0"
description = "Autonomous AI Governance Scout"
authors = [
    {name = "ChetanPatil201",email = "chetanp201jb@gmail.com"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "azure-identity (>=1.25.1,<2.0.0)",
    "azure-mgmt-compute (>=37.1.0,<38.0.0)",
    "azure-mgmt-cognitiveservices (>=14.1.0,<15.0.0)",
    "azure-mgmt-storage (>=23.1.0,<24.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "langchain-openai (>=1.1.6,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pygithub (>=2.8.1,<3.0.0)"
]


[tool.poetry]
packages = [{ include = "aigovern_warden", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
